Airflow/requirements.txt
pandas 
Faker
google-cloud-bigquery
google-auth